feat: Difference 클래스 생성 및 상속 관계 설정
